requirements.txt
camelot-py[cv]
pandas
numpy
opencv-python-headless
pdfminer.six
SQLAlchemy
psycopg2-binary
python-dotenv
ollama
streamlit
plotly
thefuzz

# Level 2: RAG with LlamaIndex (Gemini via google-genai)
llama-index
google-genai
